requirements.txt
dj-database-url==0.5.0
Django==2.2.1
django-appconf==1.0.3
django-compressor==2.3
django-compressor-toolkit==0.6.0
django-filter==2.1.0
django-guardian==1.5.1
django-heroku==0.3.1
django-webpack-loader==0.6.0
djangorestframework==3.9.4
gunicorn==19.9.0
Markdown==3.1.1
psycopg2==2.8.2
pytz==2019.1
rcssmin==1.0.6
rjsmin==1.1.0
six==1.12.0
sqlparse==0.3.0
whitenoise==4.1.2
